Minor modifications to readme
Just some slight changes to the readme to keep it similar to some of the other cores with linkbacks to explanations of what MiSTer is, short description which helps google search optimization with the word FPGA combined. Also provided linkbacks to the source for the F18A and TMS99095 implementations.

Thank you so much for the work on this core!